pyproject.toml
[project]
name = "faifwm-backend"
version = "0.1.0"
description = "FAIFWM Backend - Distributed ML Engine for Audio Protection"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.100.0",
    "uvicorn>=0.23.0",
    "python-multipart>=0.0.6",
    "grpcio>=1.59.0",
    "grpcio-tools>=1.59.0",
    "PyJWT>=2.8.0",
    "supabase>=2.3.0",
    # PyTorch dependencies need to be installed specifically for CUDA 12.8 support via UV index
    "torch>=2.2.0",
    "torchaudio>=2.2.0"
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.1",
    "ruff>=0.3.0",
    "mypy>=1.5.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
line-length = 88
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I"] # Basic PEP8, Pyflakes, isort

[tool.mypy]
strict = true
ignore_missing_imports = true
